+++ tests/fixtures/config-project/.helix/nrepl-jack-in.scm
@@ -1,0 +1,3 @@
+(nrepl-configure-jack-in 'babashka
+  (lambda (port) (string-append "custom-bb " (number->string port))))
+(nrepl-set-jack-in-env '(("CFG" . "yes")))